+++ packages/insomnia/src/ui/components/list-group/unit-test-result-timestamp.tsx
@@ -1,7 +1,7 @@
 import React, { FunctionComponent } from 'react';
 import styled from 'styled-components';
 
-import { SvgIcon } from '../../components/svg-icon';
+import { SvgIcon } from '../svg-icon';
 
 export interface UnitTestResultTimestampProps {
   timeMs: String;
