+++ packages/insomnia/src/ui/components/proto-file/proto-file-list.tsx
@@ -2,8 +2,8 @@ import React, { FunctionComponent } from 'react';
 
 import { ProtoDirectory } from '../../../models/proto-directory';
 import type { ProtoFile } from '../../../models/proto-file';
-import { ListGroup, ListGroupItem } from '../../insomnia-components';
 import type { ExpandedProtoDirectory } from '../../redux/proto-selectors';
+import { ListGroup, ListGroupItem } from '../list-group';
 import { ProtoDirectoryListItem } from './proto-directory-list-item';
 import { ProtoFileListItem } from './proto-file-list-item';
