+++ packages/insomnia/src/ui/insomnia-components/index.ts
@@ -1,3 +1,2 @@
 export * from './button';
 export * from './dropdown';
-export * from './list-group';
